modified modeling_finetune.py to make pv_log and pv_pred float32 even when deepspeed changes model weights to float16

+++ models/VideoMAEv2/engine_for_finetuning.py
@@ -22,7 +22,7 @@ import utils
 
 def train_class_batch(model, samples, pv, target, criterion):
     outputs = model(samples, pv)
-    loss = criterion(outputs.squeeze(-1), target)
+    loss = criterion(outputs.float().squeeze(-1), target)
     return loss, outputs
